chore: refactor notifications from user store into separate user notifications
