+++ src/stores/User/notifications.store.ts
@@ -1,0 +1,142 @@
+import type {
+  INotification,
+  IUser,
+  NotificationType,
+} from 'src/models/user.models'
+import { action, makeObservable, toJS } from 'mobx';
+import { randomID } from 'src/utils/helpers';
+
+import { ModuleStore } from '../common/module.store';
+import { COLLECTION_NAME as USER_COLLECTION_NAME } from './user.store';
+
+import type { RootStore } from '..'
+import type { IUserPP } from 'src/models/user_pp.models'
+
+// const COLLECTION_NAME = 'user_notifications'
+
+
+export class UserNotificationsStore extends ModuleStore {
+
+  constructor(rootStore: RootStore) {
+    super(rootStore)
+    makeObservable(this)
+  }
+  
+  get user() {
+    return this.userStore.user
+  }
+
+  public getUserNotifications() {
+    return (
+      this.user?.notifications
+        ?.filter((notification) => !notification.read)
+        .sort(
+          (a, b) =>
+            new Date(b._created).getTime() - new Date(a._created).getTime(),
+        ) || []
+    )
+  }
+
+  @action
+  public async triggerNotification(
+    type: NotificationType,
+    username: string,
+    relevantUrl: string,
+  ) {
+    try {
+      const triggeredBy = this.activeUser
+      if (triggeredBy) {
+        // do not get notified when you're the one making a new comment or how-to useful vote
+        if (triggeredBy.userName === username) {
+          return
+        }
+        const newNotification: INotification = {
+          _id: randomID(),
+          _created: new Date().toISOString(),
+          triggeredBy: {
+            displayName: triggeredBy.displayName,
+            userId: triggeredBy._id,
+          },
+          relevantUrl: relevantUrl,
+          type: type,
+          read: false,
+          notified: false,
+        }
+
+        const lookup = await this.db
+          .collection<IUserPP>(USER_COLLECTION_NAME)
+          .getWhere('userName', '==', username)
+
+        const user = lookup[0]
+
+        const updatedUser: IUser = {
+          ...toJS(user),
+          notifications: user.notifications
+            ? [...toJS(user.notifications), newNotification]
+            : [newNotification],
+        }
+
+        const dbRef = this.db
+          .collection<IUser>(USER_COLLECTION_NAME)
+          .doc(updatedUser._authID)
+
+        await dbRef.set(updatedUser)
+      }
+    } catch (err) {
+      console.error(err)
+      throw new Error(err)
+    }
+  }
+
+  @action
+  public async markAllNotificationsRead() {
+    try {
+      const user = this.activeUser
+      if (user) {
+        const notifications = toJS(user.notifications)
+        notifications?.forEach((notification) => (notification.read = true))
+        const updatedUser: IUser = {
+          ...toJS(user),
+          notifications,
+        }
+
+        const dbRef = this.db
+          .collection<IUser>(USER_COLLECTION_NAME)
+          .doc(updatedUser._authID)
+
+        await dbRef.set(updatedUser)
+        await this.userStore.updateUserProfile({ notifications })
+      }
+    } catch (err) {
+      console.error(err)
+      throw new Error(err)
+    }
+  }
+
+  @action
+  public async deleteNotification(id: string) {
+    try {
+      const user = this.activeUser
+      if (id && user && user.notifications) {
+        const notifications = toJS(user.notifications).filter(
+          (notification) => !(notification._id === id),
+        )
+
+        const updatedUser: IUser = {
+          ...toJS(user),
+          notifications,
+        }
+
+        const dbRef = this.db
+          .collection<IUser>(USER_COLLECTION_NAME)
+          .doc(updatedUser._authID)
+
+        await dbRef.set(updatedUser)
+        //TODO: ensure current user is updated
+      }
+    } catch (err) {
+      console.error(err)
+      throw new Error(err)
+    }
+  }
+}
